audit_html/requirements.txt
# Audit Intelligence — Streamlit Version
streamlit>=1.35.0
plotly>=5.18.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
openpyxl>=3.1.0
xlrd>=2.0.1
mysql-connector-python>=8.3.0
fpdf2>=2.7.0
python-multipart>=0.0.9
